refactor: rename application store to system
