+++ src/extension/apps/main/App.tsx
@@ -20,7 +20,7 @@ import { reducer as accountsReducer } from '@extension/features/accounts';
 import {
   reducer as applicationReducer,
   setSideBar,
-} from '@extension/features/application';
+} from '@extension/features/system';
 import { reducer as assetsReducer } from '@extension/features/assets';
 import { reducer as messagesReducer } from '@extension/features/messages';
 import { reducer as networksReducer } from '@extension/features/networks';
@@ -83,7 +83,7 @@ const App: FC<IAppProps> = ({ i18next, initialColorMode }: IAppProps) => {
   const store: Store<IMainRootState> = makeStore<IMainRootState>(
     combineReducers({
       accounts: accountsReducer,
-      application: applicationReducer,
+      system: applicationReducer,
       assets: assetsReducer,
       messages: messagesReducer,
       networks: networksReducer,
