+++ src/extension/components/SignTxnsModal/SignTxnsModal.tsx
@@ -41,7 +41,7 @@ import {
 } from '@common/errors';
 
 // Features
-import { setError } from '@extension/features/application';
+import { setError } from '@extension/features/system';
 import { sendSignTxnsResponse } from '@extension/features/messages';
 
 // Hooks
