+++ src/extension/enums/StoreNameEnum.ts
@@ -1,12 +1,12 @@
 enum StoreNameEnum {
   Accounts = 'accounts',
-  Application = 'application',
   Assets = 'assets',
   Messages = 'messages',
   Networks = 'networks',
   Register = 'register',
   Sessions = 'sessions',
   Settings = 'settings',
+  System = 'system',
 }
 
 export default StoreNameEnum;
