--- src/extension/features/application/types/index.ts
@@ -1,2 +1,0 @@
-export type { default as IApplicationState } from './IApplicationState';
-export type { default as IConfirm } from './IConfirm';
