+++ src/extension/features/assets/thunks/fetchAssetsThunk.ts
@@ -24,7 +24,7 @@ const fetchAssetsThunk: AsyncThunk<
   undefined,
   { state: IMainRootState }
 >(AssetsThunkEnum.FetchAssets, async (_, { getState }) => {
-  const logger: ILogger = getState().application.logger;
+  const logger: ILogger = getState().system.logger;
   const networks: INetwork[] = getState().networks.items;
   const storageManager: StorageManager = new StorageManager();
   const assetItems: Record<string, IAsset[]> = {};
