+++ src/extension/features/system/types/index.ts
@@ -1,0 +1,2 @@
+export type { default as ISystemState } from './ISystemState';
+export type { default as IConfirm } from './IConfirm';
