+++ src/extension/selectors/useSelectIsOnline.ts
@@ -4,7 +4,5 @@ import { useSelector } from 'react-redux';
 import { IMainRootState } from '@extension/types';
 
 export default function useSelectIsOnline(): boolean {
-  return useSelector<IMainRootState, boolean>(
-    (state) => state.application.online
-  );
+  return useSelector<IMainRootState, boolean>((state) => state.system.online);
 }
