+++ src/extension/selectors/useSelectLogger.ts
@@ -5,7 +5,5 @@ import { ILogger } from '@common/types';
 import { IBaseRootState } from '@extension/types';
 
 export default function useSelectLogger(): ILogger {
-  return useSelector<IBaseRootState, ILogger>(
-    (state) => state.application.logger
-  );
+  return useSelector<IBaseRootState, ILogger>((state) => state.system.logger);
 }
