+++ tsconfig.json
@@ -29,7 +29,7 @@
       "@extension/enums": ["src/extension/enums"],
       "@extension/errors": ["src/extension/errors"],
       "@extension/features/accounts": ["src/extension/features/accounts"],
-      "@extension/features/application": ["src/extension/features/application"],
+      "@extension/features/system": ["src/extension/features/system"],
       "@extension/features/assets": ["src/extension/features/assets"],
       "@extension/features/messages": ["src/extension/features/messages"],
       "@extension/features/networks": ["src/extension/features/networks"],
